requirements.txt
pywebview
tkinterweb
pystray
py2app